refactor: modified login page to popup
