--- app/src/main/java/com/taam/collection_management_system/AdminFragment.java
@@ -1,50 +1,0 @@
-package com.taam.collection_management_system;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Spinner;
-import android.widget.Toast;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-public class AdminFragment extends Fragment {
-    private EditText editTextUsername,editTextPassword;
-    private Button buttonLogin;
-
-    private FirebaseDatabase db;
-    private DatabaseReference itemsRef;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_admin_screen, container, false);
-
-        editTextPassword = view.findViewById(R.id.editTextPassword);
-        editTextUsername = view.findViewById(R.id.editTextUsername);
-        buttonLogin = view.findViewById(R.id.buttonLogin); // Initialize button here
-
-        db = FirebaseDatabase.getInstance("https://b07-demo-summer-2024-default-rtdb.firebaseio.com/");
-
-        buttonLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                adminLogin();
-            }
-        });
-
-        return view;
-    }
-
-    private void adminLogin() {
-        return;
-    }
-}
