+++ app/src/main/java/com/taam/collection_management_system/HomeFragment.java
@@ -1,6 +1,7 @@
 package com.taam.collection_management_system;
 
 import android.os.Bundle;
+import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -51,7 +52,8 @@ public class HomeFragment extends Fragment {
         buttonAdmin.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                loadFragment(new AdminFragment());
+                Intent intent = new Intent(getActivity(), AdminPopupActivity.class);
+                startActivity(intent);
             }
         });
