+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,8 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <!-- Application Theme Colors - View Importance in styles.xml   -->
+    <color name="colorPrimary">#EE8B00</color>
+    <color name="colorPrimaryDarker">#FC650D</color>
+    <color name="colorAccent">#03A1DA</color>
 </resources>
