+++ app/src/main/res/values/styles.xml
@@ -1,0 +1,9 @@
+<!-- res/values/styles.xml -->
+<resources>
+
+    <!-- Popup Activity Theme Only -->
+    <style name="AppTheme.Popup" parent="Theme.AppCompat.Dialog">
+        <item name="android:windowMinWidthMajor">90%</item>
+        <item name="android:windowMinWidthMinor">90%</item>
+    </style>
+</resources>
